Wip, made the SwingWorker instance local

+++ src/com/jetx/model/DoMath.java
@@ -2,7 +2,7 @@ package com.jetx.model;
 
 public class DoMath {
 
-	public static int fib(int number) {
+	public static long fib(long number) {
 
 		if (number < 2)
 			return number;
